refactor(frontend): move input to elements folder
